Factorize duplicated code
